20211120 se realizan cambios en el modelo de usuarios
