+++ models/usuario/usuario.js
@@ -1,16 +1,17 @@
-import { Schema, model } from 'mongoose';
-import { Enum_Rol, Enum_EstadoUsuario } from '../enums/enums';
+import mongoose from 'mongoose';
+const { Schema, model } = mongoose;
+// import { Enum_Rol, Enum_EstadoUsuario } from '../enums/enums';
 
-interface User {
-  correo: string;
-  identificacion: string;
-  nombre: string;
-  apellido: string;
-  rol: Enum_Rol;
-  estado: Enum_EstadoUsuario;
-}
+// interface User {
+//   correo: string;
+//   identificacion: string;
+//   nombre: string;
+//   apellido: string;
+//   rol: Enum_Rol;
+//   estado: Enum_EstadoUsuario;
+// }
 
-const userSchema = new Schema<User>({
+const userSchema = new Schema({
   correo: {
     type: String,
     required: true,
@@ -45,12 +46,12 @@ const userSchema = new Schema<User>({
   rol: {
     type: String,
     required: true,
-    enum: Enum_Rol,
+    enum: ['ESTUDIANTE', 'LIDER', 'ADMINISTRADOR'],
   },
   estado: {
     type: String,
-    enum: Enum_EstadoUsuario,
-    default: Enum_EstadoUsuario.PENDIENTE,
+    enum: ['PENDIENTE', 'AUTORIZADO', 'NO_AUTORIZADO'],
+    default: 'PENDIENTE',
   },
 });
